+++ CHANGELOG.md
@@ -6,7 +6,10 @@ This file lists the changes to dReams repo with each version.
 
 ### Added
 * dSkullz Collection
-* ParseSmartContract() when minting NFA
+* ParseSmartContract() for directory and when minting NFA
+* Auction highest bidder display
+* `rpc` GetNameToAddress() and send to name
+* `dreams` GetMaxSize()
 
 
 ## 0.11.1 - January 19 2024